requirements.txt
torch==2.1.*
torchvision
tensorboardX

opencv-python==4.9.*
scikit-image
monai==1.*

hydra-core
hydra-colorlog

tqdm
rootutils
pre-commit

onnx
onnxruntime
albumentations==1.3.0
mmcv-full==1.6.2
